[Feat] 명령 서버 WebSocket 구현 : S12P31S110-60

+++ BE/flip/src/main/java/com/ssafy/flip/domain/node/entity/VirtualWall.java
@@ -15,7 +15,7 @@ public class VirtualWall {
     @Id
     @Column(name = "wall_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer wallId;
+    private Long wallId;
 
     @Column(nullable = false)
     private Float x1;
